Rename "balance" expander to "node-count-expander" to better express its functionality
