+++ cluster-autoscaler/expander/factory/expander_factory.go
@@ -18,8 +18,8 @@ package factory
 
 import (
 	"k8s.io/autoscaler/cluster-autoscaler/expander"
-	"k8s.io/autoscaler/cluster-autoscaler/expander/balance"
 	"k8s.io/autoscaler/cluster-autoscaler/expander/mostpods"
+	"k8s.io/autoscaler/cluster-autoscaler/expander/nodecountbalance"
 	"k8s.io/autoscaler/cluster-autoscaler/expander/random"
 	"k8s.io/autoscaler/cluster-autoscaler/expander/waste"
 )
@@ -29,8 +29,8 @@ func ExpanderStrategyFromString(expanderFlag string) expander.Strategy {
 	var expanderStrategy expander.Strategy
 	{
 		switch expanderFlag {
-		case expander.BalanceExpanderName:
-			expanderStrategy = balance.NewStrategy()
+		case expander.NodeCountBalanceExpanderName:
+			expanderStrategy = nodecountbalance.NewStrategy()
 		case expander.RandomExpanderName:
 			expanderStrategy = random.NewStrategy()
 		case expander.MostPodsExpanderName:
